Sample a new LNA path via elliptical slice sampling.

Sample a new LNA path via elliptical slice sampling.

lna_update(
  path,
  dat,
  iter,
  parmat,
  lna_ess_schedule,
  lna_ess_control,
  initdist_objects,
  tparam,
  pathmat_prop,
  censusmat,
  draws_prop,
  ess_draws_prop,
  emitmat,
  flow_matrix,
  stoich_matrix,
  census_times,
  forcing_inds,
  forcing_tcov_inds,
  forcings_out,
  forcing_transfers,
  param_vec,
  param_inds,
  const_inds,
  tcovar_inds,
  initdist_inds,
  param_update_inds,
  census_indices,
  event_inds,
  measproc_indmat,
  svd_d,
  svd_U,
  svd_V,
  proc_pointer,
  set_pars_pointer,
  d_meas_pointer,
  do_prevalence,
  joint_initdist_update,
  step_size
)

Arguments

path
path list

dat
data matrix

iter
MCMC iteration

lna_ess_schedule

tparam
list of time-varying parameters

pathmat_prop
matrix for proposed paths

censusmat
census matrix

emitmat
emission matrix

flow_matrix
flow matrix

stoich_matrix
stoichiometry matrix

forcing_inds
forcing indices

forcing_tcov_inds
time varying covariate forcings

forcings_out
matrix with outflow from forcings

forcing_transfers
transfer matrix

param_vec
parameter vector for use in emission distribution

initdist_inds
indices of initial compartment counts

census_indices
indices where the path should be censused

event_inds
event indices

measproc_indmat
measurement process indices

svd_d
SVD objects for LNA, NULL if using the ODE approx

svd_U
SVD objects for LNA, NULL if using the ODE approx

svd_V
SVD objects for LNA, NULL if using the ODE approx

proc_pointer
C++ pointer for latent process

d_meas_pointer
C++ pointer for emission distribution

do_prevalence
should prevalence be computed

step_size
initial step size for ODE solvers

Value

list with an updated LNA path along with its stochastic perturbations, the observed dat log-likelihood, and the lna log-likelihood, and a record of the number of elliptical slice sampling proposals