Specify an MCMC transition kernel

Specify an MCMC transition kernel

mcmc_kernel(
  parameter_blocks,
  lna_ess_control = NULL,
  initdist_ess_control = NULL,
  tparam_ess_control = NULL
)

Arguments

parameter_blocks

list of parameter blocks with an MCMC kernel for each block. Each list element should be generated by a call to the parblock function

lna_ess_control

list of control settings for updating LNA paths via elliptical slice sampling, generated by a call to lna_control.

initdist_ess_control

list of control settings for updating initial distributions, generated by a call to initdist_control.

tparam_ess_control

list of control settings for updating standard normal variates for time-varying parameters via elliptical slice sampling.

Value

list containing the MCMC kernel.

Details

Specifies the MCMC transition kernel for updating parameters, the latent epidemic process, time-varying parameters, and adaptive MCMC control. Parameter blocks group (time-homogeneous) parameters that are to be updated jointly. Each parameter block is updated via a multivariate Metropolis-Hastings update or a multivariate normal slice sampling update. Empirical covariances for these algorithms can be adapted using a global adaptive scaling algorithm (Andrieu and Thoms, 2008).

References

Andrieu, Christophe, and Johannes Thoms. "A tutorial on adaptive MCMC." Statistics and Computing 18.4 (2008): 343-373.