Plot the adaptation schedule for an adaptive MCMC algorithm for a given number of iterations, given as max(1, scale_constant/(1 + iteration * step_size)^scale_cooling)

Plot the adaptation schedule for an adaptive MCMC algorithm for a given number of iterations, given as max(1, scale_constant/(1 + iteration * step_size)^scale_cooling)

plot_adaptations(
  iterations,
  scale_cooling,
  scale_constant = 1,
  step_size = 1,
  adaptation_offset = 0,
  plot_limits = c(0, 1)
)

Arguments

iterations
number of MCMC iterations

scale_cooling
rate at which to cool the adaptation, must be between 0.5 and 1.

scale_constant
maximum scaling.

step_size
increment per iteration

Value

produce a plot of the cooling schedule